revert: 드래그를 #67 의 백그라운드 구현으로 되돌려요
CallLepusMethod 빈도 경고의 원인은 예제의 @lynx-js/react/debug 가 부르는
getUniqueIdListBySnapshotId 였어요. 드래그와 무관해서 main-thread 워클릿 드래그를 걷어내요.
바텀시트의 제목 제거와 핸들 스페이서는 그대로 둬요.
