+++ package/src/components/FloatingButton.tsx
@@ -34,8 +34,10 @@ export const FloatingButton = ({
   initialPosition,
 }: FloatingButtonProps) => {
   const colors = useThemeColors();
-  const { phase, positionStyle, handlers, dragOverlayHandlers, shineRef } =
-    useDrag(bindtap, { initialPosition });
+  const { phase, positionStyle, handlers, dragOverlayHandlers } = useDrag(
+    bindtap,
+    { initialPosition },
+  );
 
   const handleReload = () => {
     try {
@@ -74,11 +76,7 @@ export const FloatingButton = ({
           style={{ backgroundColor: colors.palette.green600 }}
         >
           {children}
-          <view
-            className={"fb-shineOverlay"}
-            main-thread:ref={shineRef}
-            style={SHINE_STYLES[phase]}
-          />
+          <view className={"fb-shineOverlay"} style={SHINE_STYLES[phase]} />
         </view>
         <view
           className={"fb-reloadButton"}
